requirement.txt
django-safedelete == 1.3.1
python-dateutil == 1.4
psycopg2-binary == 2.9.5
django-ninja == 0.20.0
Django==4.1.4

